Fix #11: Added modifier_field_inverse_of on track_history that defines the inverse_of reltionship of the modifier class.
