+++ spec/integration/multi_relation_spec.rb
@@ -1,12 +1,7 @@
-require File.expand_path(File.dirname(__FILE__) + '/spec_helper')
+require 'spec_helper'
 
 describe Mongoid::History::Tracker do
-  before :each do
-    
-    class Tracker
-      include Mongoid::History::Tracker
-    end
-    
+  before :each do    
     class Model
       include Mongoid::Document
       include Mongoid::History::Trackable
@@ -14,9 +9,10 @@ describe Mongoid::History::Tracker do
       field :name, type: String
       belongs_to :user, inverse_of: :models
       has_and_belongs_to_many :external_users, class_name: "User", inverse_of: :external_models
-      
+
       track_history   :on => :name,       # track title and body fields only, default is :all
                   :modifier_field => :modifier, # adds "referenced_in :modifier" to track who made the change, default is :modifier
+                  :modifier_field_inverse_of => nil, # no inverse modifier relationship
                   :version_field => :version,   # adds "field :version, :type => Integer" to track current version, default is :version
                   :track_create   =>  false,    # track document creation, default is false
                   :track_update   =>  true,     # track document updates, default is true
@@ -28,12 +24,6 @@ describe Mongoid::History::Tracker do
       has_many :models, :dependent => :destroy, inverse_of: :user
       has_and_belongs_to_many :external_model, class_name: "Model", inverse_of: :external_users
     end
-    
-    Mongoid::History.tracker_class_name == :tracker
-  end
-
-  after :each do
-    Mongoid::History.tracker_class_name = nil
   end
 
   it "should be possible to undo when having multiple relations to modifier class" do
@@ -43,10 +33,10 @@ describe Mongoid::History::Tracker do
     model = Model.new
     model.name = "Foo"
     model.user = user
-    model.save
+    model.save!
     
     model.name = "Bar"
-    model.save
+    model.save!
     
     model.undo! user
   end
