+++ spec/tracker_spec.rb
@@ -1,17 +1,10 @@
 require File.expand_path(File.dirname(__FILE__) + '/spec_helper')
 
 describe Mongoid::History::Tracker do
-  before :each do
+  it "should set tracker_class_name when included" do
     class MyTracker
       include Mongoid::History::Tracker
     end
-  end
-
-  after :each do
-    Mongoid::History.tracker_class_name = nil
-  end
-
-  it "should set tracker_class_name when included" do
     Mongoid::History.tracker_class_name.should == :my_tracker
   end
 end
